Update price precision

+++ src/aggregation/median.rs
@@ -2,7 +2,7 @@ use super::statistics::{quantile, Rational};
 use sp_std::vec::Vec;
 
 /// It will check for underflow/overflow and list size limitations (>0) and return None in that cases.
-pub fn calculate_median(mut prices: Vec<u32>) -> Option<u32> {
+pub fn calculate_median(mut prices: Vec<u64>) -> Option<u64> {
     match prices.len() {
         0 => None,
         1 => Some(prices[0]),
